Improve accuracy when multiple path entries match
It's possible to write a spec with endpoints such that a given endpoint
path might match two or more of the spec's path entries. In the example
spec, I've added an `/animals/search` endpoint in addition to
`/animals/{id}`. One supports POST, the other GET.

These SHOULD be unambiguous to `specification.endpoint`, which takes
both a path and a method. However, it makes the routing decision
entirely based on the path and then only looks for a method to confirm
it found a match.

This leads to nil answers from queries like
`specification.endpoint('/animals/search', 'POST')`, that should return
valid data. By switching from a detect to a select, and examining all
potentially valid paths for the supported method, I eliminate this
false negative.