Bring header validation into line with spec
The OpenAPI specification says that the Headers object on a response
represents headers that *CAN* be specified with the response. The parser
treats header entries as required, which is not inline with the
specification.

This patch relaxes the behavior, and bumps the version of the Gem to
reflect the fact that it may be a breaking change for some workflows.
Tests that previously assumed headers were required have been modified
to reflect the new behavior.